+++ sandbox/preact-ts/src/app.tsx
@@ -3,7 +3,7 @@ import { card } from '../styled-system/recipes'
 import { styled } from '../styled-system/jsx'
 import { splitCssProps } from '../styled-system/jsx/is-valid-prop'
 import { HTMLStyledProps } from '../styled-system/types'
-// import '../styled-system/styles.css'
+import { grid } from '../styled-system/patterns/grid'
 
 const button = sva({
   slots: ['label', 'icon'],
@@ -80,6 +80,8 @@ export function App() {
       <SplitComponent w="2" bgColor="red.300" onClick={() => console.log('123')}>
         Click me
       </SplitComponent>
+      <div className={grid({ minChildWidth: '80px', gap: 8 })}>grid.minChildWidth 80px</div>
+      <div className={grid({ minChildWidth: '20', gap: 8 })}>grid.minChildWidth 20</div>
     </div>
   )
 }