+++ packages/token-dictionary/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @pandacss/token-dictionary
 
+## 0.19.0
+
+### Patch Changes
+
+- Updated dependencies [61831040]
+- Updated dependencies [89f86923]
+  - @pandacss/types@0.19.0
+  - @pandacss/shared@0.19.0
+
+## 0.18.3
+
+### Patch Changes
+
+- @pandacss/shared@0.18.3
+- @pandacss/types@0.18.3
+
+## 0.18.2
+
+### Patch Changes
+
+- @pandacss/shared@0.18.2
+- @pandacss/types@0.18.2
+
+## 0.18.1
+
+### Patch Changes
+
+- 566fd28a: Fix issue where virtual color does not apply DEFAULT color in palette
+- 43bfa510: Fix issue where composite tokens (shadows, border, etc) generated incorrect css when using the object syntax
+  in semantic tokens.
+  - @pandacss/shared@0.18.1
+  - @pandacss/types@0.18.1
+
+## 0.18.0
+
+### Patch Changes
+
+- Updated dependencies [ba9e32fa]
+  - @pandacss/shared@0.18.0
+  - @pandacss/types@0.18.0
+
 ## 0.17.5
 
 ### Patch Changes